+++ pyobjc-framework-Accessibility/Modules/_Accessibility_protocols.m
@@ -17,4 +17,8 @@ static void __attribute__((__used__)) use_protocols(void)
     p = PyObjC_IdToPython(@protocol(AXDataAxisDescriptor));
     Py_XDECREF(p);
 #endif
+#if PyObjC_BUILD_RELEASE >= 1201
+    p = PyObjC_IdToPython(@protocol(AXBrailleMapRenderer));
+    Py_XDECREF(p);
+#endif
 }