+++ pyobjc-framework-CallKit/PyObjCTest/test_cxaction.py
@@ -1,7 +1,0 @@
-from PyObjCTools.TestSupport import TestCase
-import CallKit
-
-
-class TestCXAction(TestCase):
-    def test_methods(self):
-        self.assertResultIsBOOL(CallKit.CXAction.isComplete)