+++ pyobjc-framework-Intents/PyObjCTest/test_infile.py
@@ -1,0 +1,8 @@
+from PyObjCTools.TestSupport import TestCase
+import Intents
+
+
+class TestINFile(TestCase):
+    def test_methods(self):
+        self.assertResultIsBOOL(Intents.INFile.removedOnCompletion)
+        self.assertArgIsBOOL(Intents.INFile.setRemovedOnCompletion_, 0)