+++ pyobjc-framework-MetalPerformanceShadersGraph/PyObjCTest/test_mpsgraphrandomops.py
@@ -1,0 +1,23 @@
+from PyObjCTools.TestSupport import TestCase
+
+import MetalPerformanceShadersGraph
+
+
+class TestMPSGraphRandomOps(TestCase):
+    def test_constants(self):
+        self.assertEqual(
+            MetalPerformanceShadersGraph.MPSGraphRandomDistributionUniform, 0
+        )
+        self.assertEqual(
+            MetalPerformanceShadersGraph.MPSGraphRandomDistributionNormal, 1
+        )
+        self.assertEqual(
+            MetalPerformanceShadersGraph.MPSGraphRandomDistributionTruncatedNormal, 2
+        )
+
+        self.assertEqual(
+            MetalPerformanceShadersGraph.MPSGraphRandomNormalSamplingInvCDF, 0
+        )
+        self.assertEqual(
+            MetalPerformanceShadersGraph.MPSGraphRandomNormalSamplingBoxMuller, 1
+        )